fix: Move NonEmptyString to data_types2
