+++ write_buffer/src/codec.rs
@@ -1,7 +1,7 @@
 //! Encode/Decode for messages
 
 use crate::core::WriteBufferError;
-use data_types::{non_empty::NonEmptyString, sequence::Sequence};
+use data_types2::{NonEmptyString, Sequence};
 use dml::{DmlDelete, DmlMeta, DmlOperation, DmlWrite};
 use generated_types::{
     google::FromOptionalField,
